+++ src/Engine2D/Animation/Animation2D.cpp
@@ -8,25 +8,30 @@
 
 namespace Engine2D::Animation {
   Animation2D::Animation2D(
-    const int frameCountX, const int frameCountY, const int frameOffsetX, const int frameOffsetY, const float frameRate,
-    const bool vertical, const bool loop, const bool reverse
+    const std::string &spriteName, const int frameCountX, const int frameCountY, const int frameOffsetX, const int frameOffsetY,
+    const float frameRate, const bool vertical, const bool loop, const bool reverse
   )
     : frameCountX(frameCountX), frameCountY(frameCountY), frameOffsetX(frameOffsetX), frameOffsetY(frameOffsetY),
       currentFrame(0), loop(loop), reverse(reverse), vertical(vertical), paused(false), completed(false),
-      frameDuration(frameRate), inverseFrameDuration(1.0f / frameRate), speed(1), inverseSpeed(1), elapsedTime(0) {}
+      frameDuration(frameRate), inverseFrameDuration(1.0f / frameRate), speed(1), inverseSpeed(1), elapsedTime(0),
+      spriteName(std::move(spriteName)) {}
+
+  int Animation2D::FrameCount() const {
+    return vertical ? frameCountY : frameCountX;
+  }
 
   Animation2D Animation2D::CreateVertical(
-    const int frameCount, const int frameOffsetX, const int frameOffsetY, const float frameRate, const bool loop,
-    const bool reverse
+    const std::string &spriteName, const int frameCount, const int frameOffsetX, const int frameOffsetY, const float frameRate,
+    const bool loop, const bool reverse
   ) {
-    return Animation2D{0, frameCount, frameOffsetX, frameOffsetY, frameRate, true, loop, reverse};
+    return Animation2D{spriteName, 0, frameCount, frameOffsetX, frameOffsetY, frameRate, true, loop, reverse};
   }
 
   Animation2D Animation2D::CreateHorizontal(
-    const int frameCount, const int frameOffsetX, const int frameOffsetY, const float frameRate, const bool loop,
-    const bool reverse
+    const std::string &spriteName, const int frameCount, const int frameOffsetX, const int frameOffsetY, const float frameRate,
+    const bool loop, const bool reverse
   ) {
-    return Animation2D{frameCount, 0, frameOffsetX, frameOffsetY, frameRate, false, loop, reverse};
+    return Animation2D{spriteName, frameCount, 0, frameOffsetX, frameOffsetY, frameRate, false, loop, reverse};
   }
 
   void Animation2D::SetFrameDuration(const float frameDuration) {
@@ -45,19 +50,5 @@ namespace Engine2D::Animation {
 
   float Animation2D::GetSpeed() const {
     return speed;
-  }
-
-  void Animation2D::Pause() {
-    paused = true;
-  }
-
-  void Animation2D::Resume() {
-    paused = false;
-  }
-
-  void Animation2D::Restart() {
-    currentFrame = reverse ? 0 : (vertical ? frameCountY : frameCountX) - 1;
-    elapsedTime = 0;
-    completed = false;
   }
 }