+++ src/Engine2D/Game2D.cpp
@@ -23,7 +23,6 @@
 #include "Engine2D/Behaviour.hpp"
 #include "Engine2D/Rendering/Camera2D.hpp"
 #include "Engine/Macros/Profiling.hpp"
-#include "Engine/Rendering/ShaderPreProcessor.hpp"
 #include "Engine2D/Animation/AnimationSystem.hpp"
 
 using Engine::ResourceManager;